# openapi.yaml
openapi: 3.0.3
info:
  title: Korvo Atlas API
  description: |
    Verifiable research infrastructure for humans and AI.

    Atlas is an open-source protocol for publishing, validating, and structuring
    research into source-backed, machine-ingestable knowledge objects anchored
    on a blockchain.

    **Authentication:**
    - All `GET` requests are public (no auth required)
    - All write operations (`POST`, `PATCH`, `DELETE`) require a valid API key
    - Pass via header: `Authorization: Bearer <api-key>`
  version: 0.2.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Korvo
    url: https://github.com/korvohq/atlas
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://atlas.korvo.io
    description: Production (canonical hub)
security: []
components:
  securitySchemes:
    BearerApiKey:
      type: http
      scheme: bearer
      description: Atlas API key (format `atl_...`). Required for write operations.
  schemas:
...
    Claim:
      type: object
      required: [id, statement, confidence, sourceIds, createdAt]
      properties:
        id: { type: string, format: uuid }
        statement: { type: string, minLength: 10 }
        confidence: { type: string, enum: [low, medium, high, verified] }
        sourceIds: { type: array, items: { type: string, format: uuid }, minItems: 1 }
        questionId: { type: string, format: uuid }
        status: { type: string, enum: [draft, published, disputed, retracted] }
        tags: { type: array, items: { type: string } }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Artifact:
      type: object
      required: [id, title, type, body, claimIds, sourceIds, createdAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        type: { type: string, enum: [brief, memo, watchlist, market_map, tracker, claim_thread, source_pack, report] }
        body: { type: string }
        summary: { type: string }
        questionId: { type: string, format: uuid }
        claimIds: { type: array, items: { type: string, format: uuid }, minItems: 1 }
        sourceIds: { type: array, items: { type: string, format: uuid }, minItems: 1 }
        status: { type: string, enum: [draft, published, under_review, archived] }
        tags: { type: array, items: { type: string } }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        onChain: { type: boolean }
        contentHash: { type: string }
        ipfsCid: { type: string }
        txHash: { type: string }
        chain: { type: string, enum: [base, arbitrum, solana, polygon, local] }
        publishedToChainAt: { type: string, format: date-time }
...
    Challenge:
      type: object
      required: [id, claimId, challengerId, reason, createdAt]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        challengerId: { type: string, format: uuid }
        reason: { type: string, minLength: 10 }
        evidence: { type: string }
        status: { type: string, enum: [open, accepted, rejected, resolved] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    Endorsement:
      type: object
      required: [id, claimId, validatorId, weight, createdAt]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        validatorId: { type: string, format: uuid }
        comment: { type: string }
        weight: { type: number, minimum: 1, maximum: 5 }
        createdAt: { type: string, format: date-time }
    Revision:
      type: object
      properties:
        id: { type: string, format: uuid }
        artifactId: { type: string, format: uuid }
        version: { type: integer }
        snapshot: { type: object }
        changedBy: { type: string }
        changeNote: { type: string }
        createdAt: { type: string, format: date-time }
...
paths:
...
  /api/claims:
    get:
      summary: List all claims
      tags: [Claims]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Claim' }
    post:
      summary: Create a claim
      tags: [Claims]
      security: [{ BearerApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '201': { description: Claim created }
        '400': { description: Validation error }
        '401': { description: Missing API key }
  /api/claims/{id}:
    get:
      summary: Get claim by ID
      tags: [Claims]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Claim' }
        '404': { description: Not found }
    patch:
      summary: Update a claim
      tags: [Claims]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Updated }
        '401': { description: Missing API key }
        '404': { description: Not found }
    delete:
      summary: Delete a claim
      tags: [Claims]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }

  # ── Artifacts ──
  /api/artifacts:
    get:
      summary: List all artifacts
      tags: [Artifacts]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Artifact' }
    post:
      summary: Create an artifact
      tags: [Artifacts]
      security: [{ BearerApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        '201': { description: Artifact created }
        '400': { description: Validation error }
        '401': { description: Missing API key }
  /api/artifacts/{id}:
    get:
      summary: Get artifact by ID
      tags: [Artifacts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Artifact' }
        '404': { description: Not found }
    patch:
      summary: Update an artifact (creates a revision snapshot)
      tags: [Artifacts]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                changedBy: { type: string, description: Who made the change }
                changeNote: { type: string, description: What was changed }
      responses:
        '200': { description: Updated artifact with _revision number }
        '401': { description: Missing API key }
        '404': { description: Not found }
    delete:
      summary: Delete an artifact
      tags: [Artifacts]
      security: [{ BearerApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '404': { description: Not found }
  /api/artifacts/{id}/history:
    get:
      summary: Get revision history for an artifact
      tags: [Artifacts]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Revision history
          content:
            application/json:
              schema:
                type: object
                properties:
                  artifactId: { type: string }
                  title: { type: string }
                  totalRevisions: { type: integer }
                  revisions:
                    type: array
                    items: { $ref: '#/components/schemas/Revision' }
        '404': { description: Artifact not found }

  # ── Validators ──
...
  /api/challenges/claim/{claimId}:
    get:
      summary: Get all challenges for a specific claim
      tags: [Challenges]
      parameters:
        - name: claimId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Challenge' }

  # ── Endorsements ──
  /api/endorsements:
    get:
      summary: List all endorsements
      tags: [Endorsements]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Endorsement' }
    post:
      summary: Endorse a claim
      tags: [Endorsements]
      security: [{ BearerApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [claimId, validatorId, weight]
              properties:
                claimId: { type: string, format: uuid }
                validatorId: { type: string, format: uuid }
                comment: { type: string }
                weight: { type: number, minimum: 1, maximum: 5 }
      responses:
        '201': { description: Endorsement created }
        '400': { description: Validation error or referenced object not found }
        '401': { description: Missing API key }
        '409': { description: Validator already endorsed this claim }
  /api/endorsements/claim/{claimId}:
    get:
      summary: Get all endorsements for a claim (with summary stats)
      tags: [Endorsements]
      parameters:
        - name: claimId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Endorsements with aggregate scores
          content:
            application/json:
              schema:
                type: object
                properties:
                  claimId: { type: string }
                  endorsements:
                    type: array
                    items: { $ref: '#/components/schemas/Endorsement' }
                  summary:
                    type: object
                    properties:
                      count: { type: integer }
                      totalWeight: { type: number }
                      averageWeight: { type: number }

  # ── Publish (Blockchain) ──
...